Began proper stomping of enemies

+++ src/MarioAI/FastAndFurious.java
@@ -157,14 +157,14 @@ public class FastAndFurious extends KeyAdapter implements Agent {
 	}
 	
 	public void findPath(Environment observation) {
-		final int marioHeight = MarioMethods.getMarioHeightFromMarioMode(observation.getMarioMode());
+		final float marioHeight = MarioMethods.getMarioHeightFromMarioMode(observation.getMarioMode());
 		//long startTime = System.currentTimeMillis();
 		pathCreator.blockingFindPath(observation, world.getMarioNode(observation), world.getGoalNodes(0), marioController.getXVelocity(), enemyPredictor, marioHeight, world);
 		//System.out.println(System.currentTimeMillis() - startTime);
 	}
 	
 	public void startFindingPathFromPreviousPath(Environment observation) {
-		final int marioHeight = MarioMethods.getMarioHeightFromMarioMode(observation.getMarioMode());
+		final float marioHeight = MarioMethods.getMarioHeightFromMarioMode(observation.getMarioMode());
 		//long startTime = System.currentTimeMillis();
 		final ArrayList<DirectedEdge> path =  pathCreator.getBestPath();
 		final Node[] goalNodes = world.getGoalNodes(0);
